+++ azure-storage-blob/azure/storage/blob/appendblobservice.py
@@ -68,9 +68,9 @@ class AppendBlobService(BaseBlobService):
     '''
     MAX_BLOCK_SIZE = 4 * 1024 * 1024
 
-    def __init__(self, account_name=None, account_key=None, sas_token=None,
-                 is_emulated=False, protocol=DEFAULT_PROTOCOL, endpoint_suffix=SERVICE_HOST_BASE,
-                 custom_domain=None, request_session=None, connection_string=None, socket_timeout=None):
+    def __init__(self, account_name=None, account_key=None, sas_token=None, is_emulated=False,
+                 protocol=DEFAULT_PROTOCOL, endpoint_suffix=SERVICE_HOST_BASE, custom_domain=None, request_session=None,
+                 connection_string=None, socket_timeout=None, token_credential=None):
         '''
         :param str account_name:
             The storage account name. This is used to authenticate requests 
@@ -109,11 +109,15 @@ class AppendBlobService(BaseBlobService):
         :param int socket_timeout:
             If specified, this will override the default socket timeout. The timeout specified is in seconds.
             See DEFAULT_SOCKET_TIMEOUT in _constants.py for the default value.
+        :param token_credential:
+            A token credential used to authenticate HTTPS requests. The token value
+            should be updated before its expiration.
+        :type `~azure.storage.common.TokenCredential`
         '''
         self.blob_type = _BlobTypes.AppendBlob
         super(AppendBlobService, self).__init__(
             account_name, account_key, sas_token, is_emulated, protocol, endpoint_suffix,
-            custom_domain, request_session, connection_string, socket_timeout)
+            custom_domain, request_session, connection_string, socket_timeout, token_credential)
 
     def create_blob(self, container_name, blob_name, content_settings=None,
                     metadata=None, lease_id=None,