+++ azure-storage-common/ChangeLog.md
@@ -2,6 +2,11 @@
 
 > See [BreakingChanges](BreakingChanges.md) for a detailed list of API breaks.
 
+## Version 1.2.0rc1:
+
+- Increased default socket timeout to a more reasonable number for Python 3.5+.
+- Fixed bug where seekable streams (request body) were not being reset for retries.
+
 ## Version 1.1.0:
 
 - Support for 2017-07-29 REST version. Please see our REST API documentation and blogs for information about the related added features.